+++ app/src/components/Slider.css
@@ -1,6 +1,6 @@
 .slider {
-  flex-grow: 1;
-  padding: 10px;
+  padding: 20px;
+  width: 100%;
 }
 
 .rc-slider {
@@ -41,7 +41,7 @@
   left: 0;
   width: 100%;
   height: 100%;
-  font-size: 0.9rem;
+  font-size: 0.8rem;
 }
 
 .rc-slider-mark-text {
