+++ app/src/sections/Results.css
@@ -26,6 +26,17 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  gap: 20px;
+  gap: 10px;
   padding: 20px;
+}
+
+@media (max-width: 600px) {
+  .chart-controls {
+    flex-wrap: wrap;
+  }
+
+  .slider {
+    padding: 10px;
+    padding-bottom: 20px;
+  }
 }
